Created Schema for Server
