+++ 3/server/index.js
@@ -1,26 +1,8 @@
-const { ApolloServer, gql } = require('apollo-server');
+const { ApolloServer } = require('apollo-server');
+const typeDefs = require('./schema');
 
-// A schema is a collection of type definitions (hence "typeDefs")
-// that together define the "shape" of queries that are executed against
-// your data.
-const typeDefs = gql`
-  type Todo {
-    message: String
-    completed: Boolean 
-  }
+const server = new ApolloServer({ typeDefs });
 
-  type Query {
-    todos: [Todo]
-  }
-`;
-
-const todos = [
-  {
-    message: 'Nichts tun',
-    completed: false,
-  },
-  {
-    message: 'Essen',
-    completed: false,
-  },
-];
+server.listen().then(({ url }) => {
+  console.log(`🚀 Server ready at ${url}`);
+});
